requirements.txt
streamlit==1.23.1
numpy==1.24.3
pandas==2.0.3 
scikit-learn==1.2.2
matplotlib==3.7.1
plotly==5.9.0
seaborn==0.12.2
pillow==9.4.0